+++ lib/include/EXRSpectralImage.h
@@ -70,7 +70,7 @@ namespace SEXR
     static SpectrumType channelType(
       const std::string &channelName,
       int &              polarisationComponent,
-      float &            wavelengths_nm);
+      double &           wavelengths_nm);
 
     /**
      * Gets the channel name used in the EXR file for a specific
@@ -83,7 +83,7 @@ namespace SEXR
      * for the given Stokes component at a specific wavelength.
      */
     static std::string
-    getEmissiveChannelName(int stokesComponent, float wavelength_nm);
+    getEmissiveChannelName(int stokesComponent, double wavelength_nm);
 
     /**
      * Gets the channel name used in the EXR file for a wavelength.
@@ -93,7 +93,7 @@ namespace SEXR
      * @returns std::string containing the reflective channel name
      * at a specific wavelength.
      */
-    static std::string getReflectiveChannelName(float wavelength_nm);
+    static std::string getReflectiveChannelName(double wavelength_nm);
 
     static constexpr const char *SPECTRUM_TYPE_ATTR     = "Spectrum type";
     static constexpr const char *LENS_TRANSMISSION_ATTR = "Lens transmission";
